requirements.txt
flask
requests
beautifulsoup4
langchain_community
pinecone-client
openai
python-dotenv